requirements.txt
# BharatIntel — Dependencies
# Install: pip install -r requirements.txt

# ── HTTP & Networking ─────────────────────────────────────
httpx>=0.27.0

# ── RSS Parsing ───────────────────────────────────────────
feedparser>=6.0.0

# ── Config ────────────────────────────────────────────────
pyyaml>=6.0

# ── Logging ───────────────────────────────────────────────
structlog>=24.0.0

# ── Environment Variables ─────────────────────────────────
python-dotenv>=1.0.0

# ── LLM Integration ──────────────────────────────────────
litellm>=1.40.0

# ── Semantic Deduplication ────────────────────────────────
sentence-transformers>=3.0.0
numpy>=1.24.0
# ── PDF Generation ────────────────────────────────────────────
jinja2>=3.1.0
reportlab>=4.1.0
weasyprint>=62.0

# ── Scheduling ────────────────────────────────────────────────
apscheduler>=3.10.0

# ── Web API ───────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn>=0.29.0